backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pypdf==4.2.0
faiss-cpu==1.8.0
numpy==1.26.4
openai>=1.52.0
torch==2.3.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
sentence-transformers==2.7.0
